[mypyc] Split BinaryIntOp and introduce ComparisonOp, implements is/is not op (#9313)
BinaryIntOp used to represent arithmetic, bitwise and comparison operations on integer operations. However, this design prevents us to compare pointer types and all comparison operations should be of boolean return type while manually specifying this in BinaryIntOp is both verbose and error-prone.

This PR splits BinaryIntOp and moves the comparison functionalities to ComparsionOp.

Based on the new op, this PR also implements is and is not op.