[mypyc] Add primitives for bitwise ops (#9529)
This adds primitives for `&`, `|`, `^`, `<<`, `>>` and `~`. These speed up a
bitwise op microbenchmark by about 10x on Linux.

Long integer operations are only slightly faster, except for certain cases
of mixed short/long integers, where we are significantly faster (but only
maybe 1.5x or so).

Also support `bool` when relevant, since it's a subclass of `int` and 
overrides some operations.

Work on mypyc/mypyc#751.