Adding more information to "Too Few Arguments" (#9796)
Fixes #9340

As mentioned in the feature request, mypy now gives more information about the missing required arguments when creating an instance.
The problem was with the if statement that only gave information when the arguments provided were not empty and none. Removing that check gives information for calls with no argument too.