Add Python script to build wheels using cibuildwheel (#10096)
The contents are extracted from the current GitHub action definition:
https://github.com/mypyc/mypy_mypyc-wheels/blob/master/.github/workflows/build.yml

This is a fairly direct translation of the existing behavior. The
behavior should be identical to the current workflow.

The idea is to make this easier to maintain and easier to test
locally.

This should also make it easier to fix #10074.