Add optional CI run against python nightly (#10213)
This commit adds an optional CI run with python nightly. It is allowed to allow run failures, so 
even if the nightly run fails the whole CI still succeeds.

This can be useful for detecting breaking changes in newer python versions and tracking 
progress towards supporting them.

This adds around 17 minutes to the total time calculated by travis. Due to the fast_finish 
attribute the success of the CI can be determined, even if the nightly run isn't finished yet, 
therefore no actual runtime is added.