[mypyc] Support __setitem__, __delitem__, __len__ and __contains__ (#10451)
Previously these couldn't be used in compiled classes.

__setitem__ and __delitem__ are implemented by a single wrapper function. If the value is
NULL, it should act as __delitem__. This makes the implementation non-trivial. First, we
need to support both a single variant and two variants being defined in a class. Second, when
overriding, we need to use super(), if we only override one of them.

The other dunders are straightforward.

The implementation is pretty verbose. I'll look at refactoring it in a separate PR.

This has some overlap with the previous PR #10211 by @sohailsomani.