[mypyc] Support various number-related dunders (#10679)
This adds support for these unary dunders:
* `__neg__`
* `__invert__`
* `__int__`
* `__float__`

Also add support for binary, reversible dunders, such as `__add__` and `__radd__`.

Finally, add support for in-place operator dunders such as `__iadd__`.

The semantics of the binary dunders don't always match Python semantics, but
many common use cases should work.

There is one significant difference from Python that is not easy to remove: if a
forward dunder method is called with an incompatible argument, it's treated the
same as if it returned `NotImplemented`. This is necessary since the body of
the method is never reached on incompatible argument type and there is no
way to explicitly return `NotImplemented`. However, it's still recommended that 
the body returns `NotImplemented` as expected for Python compatibility.

If a dunder returns `NotImplemented` and has a type annotation, the return
type should be annotated as `Union[T, Any]`, where `T` is the return value 
when `NotImplemented` is not returned.

Work on mypyc/mypyc#839.