Preserve types of functions registered with singledispatch (#10756)
When calling a function that has been registered as an implementation
for a singledispatch function, currently, mypy doesn't type check
arguments to those functions when those functions are called directly
(instead of through the main singledispatch function). This fixes that by
preserving the type of the registered function even after the register decorator
has been used.