Support typing_extensions.overload (#12602)
This always existed in typing_extensions, but was an alias for
typing.overload. With python/typing#1140, it will actually make
a difference at runtime which one you use.

Note that this shouldn't change mypy's behaviour, since
we alias typing_extensions.overload to typing.overload
in typeshed, but this makes the logic less fragile.