Delete sametypes.py (#13311)
Fixes #7924

So this is not 100% refactor, I make several semantic changes. As discussed in the issue, current implementation is not very principled, so I replace old complex logic with two options:
* New `is_same_types()` is now located in `subtypes.py` and is doing two-way proper subtyping check.
* Type equality with `==` can be used when we are checking for representation equality (used rarely, mostly for internal things)

Btw this uncovered two actual bugs, one of which I fix here (two-line change), and leave a TODO for the second.